Another small thing, hopefully the last :)

+++ org/usfirst/frc/team2412/robot/sd/SmartDashboardUtils.java
@@ -60,7 +60,7 @@ public class SmartDashboardUtils {
 			while (true) {
 				Scheduler.getInstance().run();
 				
-				SmartDashboard.putString(TIMEREMAINING, (int) (DriverStation.getInstance().getMatchTime()/60) + ":" + (new String(""+ (DriverStation.getInstance().getMatchTime()%60)).length()==2 ? "" : "0") + (new String(""+ (DriverStation.getInstance().getMatchTime()%60))));
+				SmartDashboard.putString(TIMEREMAINING, (int) (DriverStation.getInstance().getMatchTime()/60) + ":" + (new String(""+ ((int) DriverStation.getInstance().getMatchTime()%60)).length()==2 ? "" : "0") + (new String(""+ (DriverStation.getInstance().getMatchTime()%60))));
 
 				
 				try {
